backend/requirements.txt
# JustTime Backend Dependencies - FastAPI + SQLAlchemy + PostgreSQL
# Archivo: requirements.txt

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# File Management
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2